+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 203
 * Update networking refresher doc
 * Include error code in call to set_option
 * saved_handler is a public interface
+* Use new saved_handler in websocket
 
 --------------------------------------------------------------------------------